requirements.txt
pandas==1.4.4
Pillow==9.4.0
plotly==5.6.0
streamlit==1.11.1
